chore: Move ItemCombination files to its own package
